transliterate/requirements.txt
Flask==2.3.2
ai4bharat-transliteration==1.1.3
pytz==2023.3
